# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAP API
  version: 1.0.0
  description: Geospatial infrastructure reporting platform API
servers:
  - url: https://sigap.live/api
    description: Production
paths:
...
  /reports/{id}/priority:
    get:
      operationId: getReportPriority
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Priority score
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: number
                  confidence:
                    type: string
                  model_version:
                    type: string
                  factors:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        value:
                          type: number
                        max:
                          type: number
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/priority-config:
    get:
      operationId: getPriorityConfig
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Priority configuration
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: updatePriorityConfig
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Priority config updated
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /audit/export:
    get:
      operationId: exportAuditLog
      tags: [audit]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Audit log export
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== OUTBOX ====================
...
  /operator/cases/{id}/priority:
    get:
      operationId: getOperatorCasePriority
      tags: [operator]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Priority info
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auditor/audit-export:
    get:
      operationId: auditorExportAudit
      tags: [auditor]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Audit export
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        details:
          type: object
          nullable: true

    # Auth
...